在分析模块中添加表存在性检查，确保查询的表名有效；在多个执行器类中添加cols()方法以获取列元数据。

+++ src/execution/executor_insert.h
@@ -58,7 +58,7 @@ class InsertExecutor : public AbstractExecutor {
             auto ih = sm_manager_->ihs_.at(sm_manager_->get_ix_manager()->get_index_name(tab_name_, index.cols)).get();
             char* key = new char[index.col_tot_len];
             int offset = 0;
-            for(size_t j = 0; j < index.col_num; ++j) {  // 使用不同的变量名j
+            for(int j = 0; j < index.col_num; ++j) {  // 使用不同的变量名j
                 memcpy(key + offset, rec.data + index.cols[j].offset, index.cols[j].len);
                 offset += index.cols[j].len;
             }
@@ -68,4 +68,5 @@ class InsertExecutor : public AbstractExecutor {
         return std::make_unique<RmRecord>(rec);
     }
     Rid &rid() override { return rid_; }
+
 };
